move functions to init
